Moved the CPU capability logic inside fipsmodule
