+++ crypto/fipsmodule/bcm.c
@@ -113,6 +113,15 @@
 #include "sha/sha256.c"
 #include "sha/sha512.c"
 #include "tls/kdf.c"
+#include "cpucap/cpu-aarch64-fuchsia.c"
+#include "cpucap/cpu-aarch64-linux.c"
+#include "cpucap/cpu-aarch64-win.c"
+#include "cpucap/cpu-arm-linux.c"
+#include "cpucap/cpu-arm.c"
+#include "cpucap/cpu-intel.c"
+#include "cpucap/cpu-ppc64le.c"
+#include "cpucap/cpucap.c"
+
 
 #if defined(BORINGSSL_FIPS)
