+++ crypto/fipsmodule/cpucap/cpu-aarch64-linux.c
@@ -30,8 +30,6 @@
 
 #include <openssl/arm_arch.h>
 
-#include "internal.h"
-
 
 extern uint32_t OPENSSL_armcap_P;
